requirements.txt
numpy
mxnet-cu80
matplotlib
seaborn